requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
colorama==0.4.4
Django==3.1.5
djangorestframework==3.12.2
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
wrapt==1.12.1
